+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ### v3000.1.7
 - fixed `debug.paused` not pausing audio
+- added `mask()` component
+- added support for colored font outline
+```js
+loadFont("apl386", "/examples/fonts/apl386.ttf", {
+	outline: {
+		width: 8,
+		color: rgb(0, 0, 255),
+	},
+})
+```
 
 ### v3000.1.6
 - fixed `loadSound` typing to accept `ArrayBuffer`