+++ CHANGELOG.md
@@ -12,6 +12,7 @@ loadFont("apl386", "/examples/fonts/apl386.ttf", {
 ```
 - fixed `wave()` not starting at `0` when time is `0`
 - kaboom now only displays error screen for kaboom's own error, instead of catching all errors in current window
+- added `KaboomError` class for errors related to current kaboom instance
 
 ### v3000.1.6
 - fixed `loadSound` typing to accept `ArrayBuffer`